dev-requirements.txt
# These are dev dependencies; CLI's and tools used for local development. Dependencies that are needed for the Lambda
# to run should be included in the production-requirements.txt

black==22.12.0
flake8==4.0.1
importlib_metadata==4.13.0
isort==5.13.1
pytest==7.4.3
pytest-dotenv==0.5.2
moto==4.2.11
freezegun==1.3.1
boto3-stubs[dynamodb]==1.33.12
flake8-print==5.0.0
aws-lambda-powertools[all]
terraform-local==0.16.1
